refactor: Restructure shortcuts management and update imports
- Moved shortcuts-related components and types to a new directory structure for better organization.
- Updated import paths across various files to reflect the new structure.
- Removed obsolete shortcut files to streamline the codebase and improve maintainability.
