+++ src/components/XPlayer/components/Shortcuts/components/ShortcutsItem.vue
@@ -36,8 +36,8 @@
 import type { Action, ActionKey, KeyBindings, KeyBindingStr } from '../shortcuts.types'
 import { Icon } from '@iconify/vue'
 import { computed } from 'vue'
-import { usePlayerContext } from '../../hooks/usePlayerProvide'
-import { ICONS } from '../../index.const'
+import { usePlayerContext } from '../../../hooks/usePlayerProvide'
+import { ICONS } from '../../../index.const'
 import {
   hasEmptyKeybindings,
   isMaxKeybindings,
