+++ src/components/XPlayer/components/Shortcuts/index.ts
