+++ src/components/XPlayer/components/Shortcuts/shortcuts.utils.ts
